Make external benches.
