+++ src/lib.rs
@@ -4,10 +4,7 @@
 #![crate_name = "event"]
 #![deny(missing_docs)]
 #![deny(missing_copy_implementations)]
-#![cfg_attr(test, feature(test))]
 
-#[cfg(test)]
-extern crate test;
 extern crate input;
 extern crate event_loop;
 extern crate window;
