+++ src/press.rs
@@ -40,7 +40,6 @@ impl<T: GenericEvent> PressEvent for T {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use test::Bencher;
 
     #[test]
     fn test_input_press() {
@@ -54,16 +53,6 @@ mod tests {
         assert_eq!(x, y);
     }
 
-    #[bench]
-    fn bench_input_press(bencher: &mut Bencher) {
-        use input::{ Button, Input, Key };
-
-        let e = Input::Press(Button::Keyboard(Key::S));
-        let button = Button::Keyboard(Key::A);
-        bencher.iter(|| {
-            let _: Option<Input> = PressEvent::from_button(button, &e);
-        });
-    }
 
     #[test]
     fn test_event_press() {
@@ -76,17 +65,5 @@ mod tests {
         let y: Option<Event> = x.clone().unwrap().press(|button|
             PressEvent::from_button(button, x.as_ref().unwrap())).unwrap();
         assert_eq!(x, y);
-    }
-
-    #[bench]
-    fn bench_event_press(bencher: &mut Bencher) {
-        use Event;
-        use input::{ Button, Key, Input };
-
-        let e = Event::Input(Input::Press(Button::Keyboard(Key::S)));
-        let button = Button::Keyboard(Key::A);
-        bencher.iter(|| {
-            let _: Option<Event> = PressEvent::from_button(button, &e);
-        });
     }
 }
